% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augmentation.R
\name{augmentation}
\alias{augmentation}
\title{Augmentation}
\usage{
augmentation(xf, xnf, yf)
}
\arguments{
\item{xf}{The matrix of financed clients' characteristics to be used in the scorecard.}

\item{xnf}{The matrix of not financed clients' characteristics to be used in the scorecard (must be the same in the same order as xf!).}

\item{yf}{The matrix of financed clients' labels}
}
\description{
This function performs Reject Inference using the Augmentation technique. Note that this technique is theoretically better than using the financed clients scorecard in the MAR and misspecified model case.
}
\examples{
augmentation()
}
\keyword{credit,}
\keyword{inference,}
\keyword{r?int?gration,}
\keyword{reject,}
\keyword{scorecard,}
\keyword{scoring}
