% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_consistency.R
\name{check_consistency}
\alias{check_consistency}
\title{Check consistency for reject inference inputs}
\usage{
check_consistency(xf, xnf, yf)
}
\arguments{
\item{xf}{The matrix of financed clients' characteristics to be used in the scorecard.}

\item{xnf}{The matrix of not financed clients' characteristics to be used in the scorecard (must be the same in the same order as xf!).}

\item{yf}{The matrix of financed clients' labels}
}
\description{
This function checks consistency for reject inference inputs
}
\examples{
\dontrun{
# We simulate data from financed clients
xf <- matrix(runif(100 * 2), nrow = 100, ncol = 2)
theta <- c(2, -2)
log_odd <- apply(xf, 1, function(row) theta \%*\% row)
yf <- rbinom(100, 1, 1 / (1 + exp(-log_odd)))
xnf <- matrix(runif(100 * 2), nrow = 100, ncol = 2)
check_consistency(xf, xnf, yf)
}
}
\seealso{
\code{glm}, \code{speedglm}
}
\author{
Adrien Ehrhardt

Adrien Ehrhardt
}
\keyword{internal}
