% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut.dataset.R
\name{cut_dataset}
\alias{cut_dataset}
\title{Randoming and sampling a dataset by its row numbers}
\usage{
cut_dataset(n, proportions, test = TRUE, validation = TRUE)
}
\arguments{
\item{n}{The total number of rows to sample from.}

\item{proportions}{The list of the (2) proportions wanted for test and validation set. Only the first is used when there is only one of either test or validation that is set to TRUE. Produces an error when the sum is greater to one. Useless if both test and validation are set to FALSE. Default: list(0.2,0.2).}

\item{test}{TRUE if a test dataset is wanted, FALSE otherwise (default=TRUE).}

\item{validation}{TRUE if a test dataset is wanted, FALSE otherwise (default=TRUE).}
}
\description{
This function chooses the rows of a dataset to be included in the train, test (optional) and validation (optional) datasets according to proportions specified by the user.
}
\author{
Adrien Ehrhardt
}
\keyword{internal}
\keyword{sample}
\keyword{test}
\keyword{train}
\keyword{validation}
