% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{discretization}
\alias{discretization}
\title{Class discretization}
\description{
An S4 class to represent a discretization scheme.
}
\section{Slots}{

\describe{
\item{\code{method.name}}{The name of the used discretization method.}

\item{\code{parameters}}{The parameters associated with the used method.}

\item{\code{best.disc}}{The best discretization scheme found by the method given its parameters.}

\item{\code{performance}}{The performance obtained with the method given its parameters.}

\item{\code{disc.data}}{The discretized data: test set if test is TRUE; if test is FALSE and validation is TRUE, then it provides the discretized validation set. Otherwise, it provides the discretized training set.}

\item{\code{disc.data}}{The continuous data: test set if test is TRUE; if test is FALSE and validation is TRUE, then it provides the discretized validation set. Otherwise, it provides the discretized training set.}
}}

