% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretize.link.R
\name{discretize.link}
\alias{discretize.link}
\title{Discretize a dataset using a trained SEM discretization scheme}
\usage{
discretize.link(link, df)
}
\arguments{
\item{link}{A multinomial logit model.}

\item{df}{The dataframe, containing the same variables as the one used to train the discretization scheme, to be discretized.}
}
\description{
This function discretizes a user-specified dataset using a pre-trained SEM discretization scheme.
}
\examples{
# discretize.link()
}
\keyword{discretization,}
\keyword{predict}
