% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\name{discretize}
\alias{discretize}
\title{Generic method "discretize" for discretization objects.}
\description{
This defines the generic method "discretize" which will discretize a new input dataset given a discretization scheme of S4 class discretization.

This defines the method "discretize" which will discretize a new input dataset given a discretization scheme of S4 class discretization.
}
