% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretize.cutp.R
\name{discretize_cutp}
\alias{discretize_cutp}
\title{Discretization of a test dataset when provided with an already discretized training set.}
\usage{
discretize_cutp(cont_test_set, disc_train_set, cont_train_set)
}
\arguments{
\item{cont_test_set}{The matrix or dataframe of the dataset to discretize.}

\item{disc_train_set}{The matrix or dataframe of the discretized training dataset.}

\item{cont_train_set}{The matrix or dataframe of the continuous training dataset.}
}
\description{
This function discretizes a dataset when provided with both continuous and discretized versions of a training dataset.
}
\author{
Adrien Ehrhardt
}
\keyword{internal}
\keyword{sample}
\keyword{test}
\keyword{train}
\keyword{validation}
