% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretize.link.R
\name{discretize_link}
\alias{discretize_link}
\title{Discretize a dataset using a trained SEM discretization scheme}
\usage{
discretize_link(link, df)
}
\arguments{
\item{link}{A multinomial logit model.}

\item{df}{The dataframe, containing the same variables as the one used to train the discretization scheme, to be discretized.}
}
\description{
This function discretizes a user-specified dataset using a pre-trained SEM discretization scheme.
}
\examples{
set.seed(1)
x = matrix(runif(100), nrow = 100, ncol = 1)
cuts = seq(0,1,length.out = 4)
xd = as.numeric(cut(x,cuts))

long_dataset <- data.frame(e = as.vector(sapply(xd,function(var) (seq(1:3)[seq(1:3)]==var))),x = as.vector(sapply(x[,1], function(var) rep(var,3))), names = as.character(as.vector(rep(seq(1:3)[seq(1:3)],length(x)))),stringsAsFactors=FALSE)
link <- mnlogit::mnlogit(e ~ 1 | x | 1, data=long_dataset, choiceVar = "names")
discretize_link(link,as.data.frame(x))
}
\references{
Asad Hasan, Wang Zhiyu and Alireza S. Mahani (2015). mnlogit: Multinomial Logit Model. R package version 1.2.4. \url{https://CRAN.R-project.org/package=mnlogit}
}
\keyword{discretization,}
\keyword{predict}
