% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_data}
\alias{generate_data}
\title{Generate data following different missingness mechanisms}
\usage{
generate_data(n = 100, d = 3, type = "MAR well specified")
}
\arguments{
\item{n}{The number of samples to return.}

\item{d}{The dimension of samples to return.}

\item{type}{The matrix of financed clients' labels}
}
\value{
Dataframe containing features as x.1..d, labels as y.
}
\description{
This function performs generates
}
\details{
This function generates data from a uniform(0,1) distribution, and generates
labels y according to a logistic regression on this data with random -1/1
parameter for each coordinate (MAR well-specified), the square of this data
(MAR misspecified), or this data and some additional feature (from U(0,1) as
well - MNAR).
}
\examples{
# We simulate data from financed clients
generate_data(n = 100, d = 3, type = "MAR well specified")
}
\references{
Ehrhardt, A., Biernacki, C., Vandewalle, V., Heinrich, P. and Beben, S. (2018), Reject Inference Methods in Credit Scoring: a rational review,
}
\author{
Adrien Ehrhardt
}
\keyword{credit}
\keyword{inference}
\keyword{reject}
\keyword{réintégration}
\keyword{scorecard}
\keyword{scoring}
