% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cutp.R
\name{get_cutp}
\alias{get_cutp}
\title{Getting cut-points of a discretization algorithm given the continuous and the discretized sets.}
\usage{
get_cutp(disc_train_set, cont_train_set)
}
\arguments{
\item{disc_train_set}{The matrix or dataframe of the discretized training dataset.}

\item{cont_train_set}{The matrix or dataframe of the continuous training dataset.}
}
\description{
This function returns the cut-points
}
\author{
Adrien Ehrhardt
}
\keyword{internal}
\keyword{sample}
\keyword{test}
\keyword{train}
\keyword{validation}
