% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lendingClub}
\alias{lendingClub}
\title{Lending Club mortgages.}
\format{A data frame with 2167 rows and 16 variables.}
\source{
\url{https://www.lendingclub.com/}
}
\usage{
lendingClub
}
\description{
A dataset containing the information about Lending Club loans available online.
}
\keyword{datasets}
