% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_f.R
\name{model_finances}
\alias{model_finances}
\title{Financed model}
\usage{
model_finances(df_f)
}
\arguments{
\item{df_f}{data frame of labels and financed clients' data}
}
\value{
Logistic regression model
}
\description{
This function learns a logistic regression on financed clients only
}
\examples{
\dontrun{
# We simulate data from financed clients
xf <- matrix(runif(100 * 2), nrow = 100, ncol = 2)
theta <- c(2, -2)
log_odd <- apply(xf, 1, function(row) theta \%*\% row)
yf <- rbinom(100, 1, 1 / (1 + exp(-log_odd)))
df_f <- data.frame(labels = yf, x = xf)
model_finances(df_f)
}
}
\references{
Enea, M. (2015), speedglm: Fitting Linear and Generalized Linear Models to Large Data Sets, \url{https://CRAN.R-project.org/package=speedglm}
Ehrhardt, A., Biernacki, C., Vandewalle, V., Heinrich, P. and Beben, S. (2018), Reject Inference Methods in Credit Scoring: a rational review,
}
\seealso{
\code{glm}, \code{speedglm}
}
\author{
Adrien Ehrhardt
}
