% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizedGini.R
\name{normalizedGini}
\alias{normalizedGini}
\title{Calculating the normalized Gini index}
\usage{
normalizedGini(actual, predicted)
}
\arguments{
\item{actual}{The numeric binary vector of the actual labels observed.}

\item{predicted}{The vector of the probabilities predicted by the classification rule.}
}
\description{
This function calculates the Gini index of a classification rule outputting probabilities.
}
\examples{
normalizedGini()
}
\keyword{gini,}
\keyword{index}
