% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\name{plot.discretization}
\alias{plot.discretization}
\title{Different kinds of plots using either plotly (if available) or the standard plot (graphics package).}
\usage{
\method{plot}{discretization}(object, type, ...)
}
\arguments{
\item{type}{The test dataframe to discretize and for which we wish to have predictions.}

\item{...}{See additional parameters of plotly (if installed) or the standard plot function (from the graphics package).}

\item{discretization}{The S4 discretization object.}
}
\description{
This function aims at producing useful graphs in the context of credit scoring in order to simplify the validation process
of the produced credit score.
}
\keyword{discretization,}
\keyword{predict,}
\keyword{prediction}
\keyword{test,}
