% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\name{plot}
\alias{plot}
\alias{plot.discretization}
\title{Different kinds of plots using either plotly (if available) or the standard plot (graphics package).}
\usage{
plot.discretization(x, type, ...)
}
\arguments{
\item{x}{S4 discretization object.}

\item{type}{Type of plot. For now only "ROC" is supported.}

\item{...}{Additional plot parameters.}
}
\description{
This function aims at producing useful graphs in the context of credit scoring in order to simplify the validation process
of the produced credit score.
}
\section{Functions}{
\itemize{
\item \code{plot.discretization}: plots for the S4 discretization object
}}

