% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\name{predict.discretization}
\alias{predict.discretization}
\title{Prediction on a raw test set of the best logistic regression model on discretized data.}
\usage{
\method{predict}{discretization}(object, predictors)
}
\arguments{
\item{discretization}{The S4 discretization object.}

\item{cont_test_set}{The test dataframe to discretize and for which we wish to have predictions.}
}
\description{
This function discretizes a user-provided test dataset given a discretization scheme provided by an S4 "discretization" object.
It then applies the learnt logistic regression model and outputs its prediction (see predict.glm).
}
\keyword{discretization,}
\keyword{predict,}
\keyword{prediction}
\keyword{test,}
