% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\name{predict}
\alias{predict}
\alias{predict.discretization}
\alias{predict.reject_infered}
\alias{predict,discretization-method}
\alias{predict,reject_infered-method}
\title{Prediction on a raw test set of the best logistic regression model on discretized data.}
\usage{
predict(object, ...)

predict.discretization(object, newdata)

predict.reject_infered(object, newdata, ...)

\S4method{predict}{discretization}(object, newdata)

\S4method{predict}{reject_infered}(object, newdata, ...)
}
\arguments{
\item{object}{The S4 discretization object.}

\item{...}{Additional parameters to pass on to base predict.}

\item{newdata}{The test dataframe to discretize and for which we wish to have predictions.}
}
\description{
This function discretizes a user-provided test dataset given a discretization scheme provided by an S4 "discretization" object.
It then applies the learnt logistic regression model and outputs its prediction (see predict.glm).
}
