% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzyAugmentation.R, R/reclassification.R
\name{reclassification}
\alias{reclassification}
\alias{reclassification}
\title{Fuzzy Augmentation}
\usage{
reclassification(xf, xnf, yf, thresh = 0.5)

reclassification(xf, xnf, yf, thresh = 0.5)
}
\arguments{
\item{xf}{The matrix of financed clients' characteristics to be used in the scorecard.}

\item{xnf}{The matrix of not financed clients' characteristics to be used in the scorecard (must be the same in the same order as xf!).}

\item{yf}{The matrix of financed clients' labels}

\item{xf}{The matrix of financed clients' characteristics to be used in the scorecard.}

\item{xnf}{The matrix of not financed clients' characteristics to be used in the scorecard (must be the same in the same order as xf!).}

\item{yf}{The matrix of financed clients' labels}
}
\description{
This function performs Reject Inference using the Fuzzy Augmentation technique. Note that this technique has no theoretical foundation and should produce (under the identifiability assumption) the same parameters' estimates than the financed clients scorecard.

This function performs Reject Inference using the Reclassification technique. Note that this technique has no theoretical foundation as it performs a one-step CEM algorithm.
}
\examples{
fuzzy()
reclassification()
}
\keyword{credit,}
\keyword{inference,}
\keyword{r?int?gration,}
\keyword{reject,}
\keyword{scorecard,}
\keyword{scoring}
