% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduireGLM.R
\name{reduireGLM}
\alias{reduireGLM}
\title{Compressing glm results}
\usage{
reduireGLM(modele)
}
\arguments{
\item{modele}{The glm model to compress.}
}
\description{
This function suppresses attributes of glm models that depend on the size of the input dataset.
}
\examples{
reduireGLM()
}
\keyword{glm,}
\keyword{size}
