% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{reject_infered-class}
\alias{reject_infered-class}
\title{Class reject_infered}
\description{
An S4 class to represent a reject inference technique.
}
\section{Slots}{

\describe{
\item{\code{method_name}}{The name of the used reject inference method.}

\item{\code{financed_model}}{The logistic regression model on financed clients.}

\item{\code{acceptance_model}}{The acceptance model (if estimated by the given method).}

\item{\code{infered_model}}{The logistic regression model resulting from the reject inference method.}
}}

