% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\docType{package}
\name{scoring-package}
\alias{scoring}
\alias{scoring-package}
\title{Generate R documentation from inline comments.}
\description{
Roxygen2 allows you to write documentation in comment blocks co-located
with code.
}
\details{
The only function you're likely to need from \pkg{scoring} is
\code{\link{scoring}}. Otherwise refer to the vignettes to see
how to format the documentation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://adimajo.github.io}
  \item Report bugs at \url{https://github.com/adimajo/scoring/issues}
}

}
