% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\docType{package}
\name{scoringTools-package}
\alias{scoringTools}
\alias{scoringTools-package}
\title{Credit Scoring Tools.}
\description{
Refer to the package's vignette.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://adimajo.github.io/scoringTools/}
  \item Report bugs at \url{https://github.com/adimajo/scoringTools/issues}
}

}
\author{
\strong{Maintainer}: Adrien Ehrhardt \email{adrien.ehrhardt@centraliens-lille.org}

}
