% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\name{summary}
\alias{summary}
\alias{summary.discretization}
\title{Summary}
\usage{
summary.discretization(object)
}
\arguments{
\item{object}{S4 discretization object.}
}
\description{
Summary generic.
}
\section{Functions}{
\itemize{
\item \code{summary.discretization}: Summary for the discretization class.
}}

