% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut.dataset.R
\name{cut.dataset}
\alias{cut.dataset}
\title{Randoming and sampling a dataset by its row numbers}
\usage{
\method{cut}{dataset}(n, test = TRUE, validation = TRUE,
  proportions = c(), seed = 1)
}
\arguments{
\item{n}{The total number of rows to sample from.}

\item{test}{TRUE if a test dataset is wanted, FALSE otherwise (default=TRUE).}

\item{validation}{TRUE if a test dataset is wanted, FALSE otherwise (default=TRUE).}

\item{proportions}{The list of the proportions.}

\item{seed}{The seed for the random number generator (optional).}
}
\description{
This function chooses the rows of a dataset to be included in the train, test (optional) and validation (optional) datasets according to proportions specified by the user.
}
\examples{
cut.dataset()
}
\keyword{sample,}
\keyword{test,}
\keyword{train,}
\keyword{validation}
