% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semDiscretization.R
\name{sem_polytomique}
\alias{sem_polytomique}
\title{Discretizing using the SEM-Gibbs algorithm}
\usage{
sem_polytomique(predictors, labels, test = TRUE, validation = TRUE,
  criterion = "gini", iter = 1000, m_depart = 20, reg_type = "poly")
}
\arguments{
\item{predictors}{The matrix or dataframe containing the .}

\item{labels}{The numeric binary vector of the actual labels observed.}

\item{test}{The numeric binary vector of the actual labels observed.}

\item{validation}{The numeric binary vector of the actual labels observed.}

\item{criterion}{The numeric binary vector of the actual labels observed.}

\item{iter}{The numeric binary vector of the actual labels observed.}

\item{m_depart}{The numeric binary vector of the actual labels observed.}

\item{reg_type}{The numeric binary vector of the actual labels observed.}
}
\description{
This function determines the best discretization for a logistic regression model using an SEM-Gibbs algorithm which affects continuous variables to qualitative variables using multinomial regression.
}
\examples{
sem_polytomique()
}
\keyword{discretization,}
\keyword{index}
