% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\name{summary}
\alias{summary}
\alias{summary.discretization}
\alias{summary,discretization-method}
\title{Summary}
\usage{
summary(object, ...)

summary.discretization(object)

\S4method{summary}{discretization}(object)
}
\arguments{
\item{object}{S4 discretization object.}

\item{...}{Other parameters to \code{summary}}
}
\description{
Summary generic.
}
