% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twins.R
\name{twins}
\alias{twins}
\title{Twins}
\usage{
twins(xf, xnf, yf)
}
\arguments{
\item{xf}{The matrix of financed clients' characteristics to be used in the scorecard.}

\item{xnf}{The matrix of not financed clients' characteristics to be used in the scorecard (must be the same in the same order as xf!).}

\item{yf}{The matrix of financed clients' labels}
}
\description{
This function performs Reject Inference using the Twins technique. Note that this technique has no theoretical foundation.
}
\examples{
twins()
}
\keyword{credit,}
\keyword{inference,}
\keyword{r?int?gration,}
\keyword{reject,}
\keyword{scorecard,}
\keyword{scoring}
